# src/public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "LinkIO API",
      "version": "1.0.0",
      "description": "API for digital dentistry",
      "contact": {
        "name": "Caio Mikoszewski",
        "email": "caio.mikoszewski@arrowdigital.group"
      },
      "license": {
        "name": "AllRights-Attribution-NonCommercial-NoDerivatives 4.1",
        "url": "https://github.com/arrow-digital/license"
      }
    },
    "servers": [
      {
        "url": "https://vastly-cheerful-cougar.ngrok-free.app",
        "description": "NGROK tunnel server"
      },
      {
        "url": "http://localhost:8080",
        "description": "Local server"
      }
    ],
    "paths": {
      "/api/models/clinic": {
        "get": {
          "summary": "Get all clinics",
          "tags": [
            "Clinic"
          ],
          "responses": {
            "200": {
              "description": "List of clinics",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Clinic"
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "post": {
          "summary": "Create a new clinic",
          "tags": [
            "Clinic"
          ],
          "requestBody": {
            "description": "Clinic object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Clinic"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Created clinic",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Clinic"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
      "/api/models/clinic/{id}": {
        "get": {
          "summary": "Get a clinic by ID",
          "tags": [
            "Clinic"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the clinic to retrieve",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Clinic object",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Clinic"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "put": {
          "summary": "Update a clinic by ID",
          "tags": [
            "Clinic"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the clinic to update",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "description": "Clinic object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Clinic"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Updated clinic",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "boolean"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "delete": {
          "summary": "Delete a clinic by ID",
          "tags": [
            "Clinic"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the clinic to delete",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Deleted clinic",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "boolean"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
      "/api/models/collaborator": {
        "get": {
            "summary": "Get all collaborators",
            "tags": [
              "Collaborator"
            ],
            "responses": {
              "200": {
                "description": "List of collaborators",
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Collaborator"
                      }
                    }
                  }
                }
              },
              "500": {
                "description": "Internal server error"
              }
            }
        },
        "post": {
            "summary": "Create a new collaborator",
            "tags": [
              "Collaborator"
            ],
            "requestBody": {
              "description": "Collaborator object",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Collaborator"
                  }
                }
              }
            },
            "responses": {
              "200": {
                "description": "Created collaborator",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/components/schemas/Collaborator"
                    }
                  }
                }
              },
              "500": {
                "description": "Internal server error"
              }
            }
        }
      },
      "/api/models/collaborator/{id}": {
        "get": {
          "summary": "Get a collaborator by ID",
          "tags": [
            "Collaborator"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the collaborator to retrieve",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Collaborator object",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Collaborator"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "put": {
          "summary": "Update a collaborator by ID",
          "tags": [
            "Collaborator"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the collaborator to update",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "description": "Collaborator object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Collaborator"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Updated collaborator",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "boolean"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "delete": {
          "summary": "Delete a collaborator by ID",
          "tags": [
            "Collaborator"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the collaborator to delete",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Deleted collaborator",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "boolean"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
      "/api/models/lab": {
        "get": {
          "summary": "Get all labs",
          "tags": [
            "Lab"
          ],
          "responses": {
            "200": {
              "description": "List of labs",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Lab"
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "post": {
          "summary": "Create a new lab",
          "tags": [
            "Lab"
          ],
          "requestBody": {
            "description": "Lab object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lab"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Created lab",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Lab"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
      "/api/models/lab/{id}": {
        "get": {
          "summary": "Get a lab by ID",
          "tags": [
            "Lab"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the lab to retrieve",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Lab object",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Lab"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "put": {
          "summary": "Update a lab by ID",
          "tags": [
            "Lab"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the lab to update",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "description": "Lab object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lab"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Updated lab",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "boolean"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "delete": {
          "summary": "Delete a lab by ID",
          "tags": [
            "Lab"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the lab to delete",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Deleted lab",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "boolean"
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
        "Clinic": {
            "properties": {
                "_id": {
                    "type": "string"
                },
                "address": {
                    "description": "must be a string and is required",
                    "type": "string"
                },
                "balance": {
                    "items": {
                        "properties": {
                            "_id": {
                                "description": "must be an objectId and is optional",
                                "type": "string"
                            },
                            "createdAt": {
                                "description": "must be a date and is required",
                                "format": "date-time",
                                "type": "string"
                            },
                            "expiresAt": {
                                "description": "must be a date and is required",
                                "format": "date-time",
                                "type": "string"
                            },
                            "order": {
                                "description": "must be an objectId and is optional",
                                "type": "string"
                            },
                            "service": {
                                "description": "must be an objectId and is optional, references the service in the order",
                                "type": "string"
                            },
                            "value": {
                                "description": "must be a number and is required",
                                "type": "number"
                            }
                        },
                        "required": [
                            "value",
                            "createdAt",
                            "expiresAt"
                        ],
                        "type": "object"
                    },
                    "type": "array"
                },
                "dentists": {
                    "description": "must be an array of objects and is required",
                    "items": {
                        "properties": {},
                        "required": [],
                        "type": "object"
                    },
                    "type": "array"
                },
                "email": {
                    "description": "must be a string and is optional",
                    "type": "string"
                },
                "name": {
                    "description": "must be a string and is required",
                    "type": "string"
                },
                "orders": {
                    "description": "must be an array of objectIds and is required",
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                },
                "outstandingBalance": {
                    "description": "must be a number and is required",
                    "type": "number"
                },
                "patients": {
                    "description": "must be an array of objects and is required",
                    "items": {
                        "properties": {},
                        "required": [],
                        "type": "object"
                    },
                    "type": "array"
                },
                "phone": {
                    "description": "must be a string and is optional",
                    "type": "string"
                },
                "taxid": {
                    "description": "must be a string and is optional",
                    "type": "string"
                }
            },
            "required": [
                "name",
                "address",
                "patients",
                "dentists",
                "orders",
                "outstandingBalance"
            ],
            "title": "Clinic"
        },
        "Collaborator": {
            "properties": {
                "_id": {
                    "type": "string"
                },
                "address": {
                    "description": "must be a string and is optional",
                    "type": "string"
                },
                "bankInfo": {
                    "description": "must be an object and is optional",
                    "properties": {},
                    "type": "object"
                },
                "comission": {
                    "description": "must be a number and is optional",
                    "type": "number"
                },
                "email": {
                    "description": "must be a string and is optional",
                    "type": "string"
                },
                "lab": {
                    "description": "must be an objectId and is optional",
                    "type": "string"
                },
                "name": {
                    "description": "must be a string and is required",
                    "type": "string"
                },
                "phone": {
                    "description": "must be a string and is optional",
                    "type": "string"
                },
                "role": {
                    "description": "must be a string and is optional",
                    "type": "string"
                },
                "salary": {
                    "description": "must be a number and is optional",
                    "type": "number"
                },
                "taxid": {
                    "description": "must be a string and is optional",
                    "type": "string"
                },
                "type": {
                    "description": "must be a string and is optional",
                    "type": "string"
                }
            },
            "required": [
                "name"
            ],
            "title": "Collaborator"
        },
...
        "Lab": {
            "properties": {
                "_id": {
                    "type": "string"
                },
                "address": {
                    "description": "must be a string and is optional",
                    "type": "string"
                },
                "clinics": {
                    "description": "must be an array of objectIds from the clinics collection and is optional",
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                },
                "collaborators": {
                    "description": "must be an array of objectIds from the users collection and is optional",
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                },
                "email": {
                    "description": "must be a string and is optional",
                    "type": "string"
                },
                "inventory": {
                    "description": "must be an array of objectIds from the inventory collection and is optional",
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                },
                "name": {
                    "description": "must be a string and is required",
                    "type": "string"
                },
                "orders": {
                    "description": "must be an array of objectIds from the orders collection and is optional",
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                },
                "phone": {
                    "description": "must be a string and is optional",
                    "type": "string"
                },
                "revenue": {
                    "description": "must be an array of objects and is required",
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                },
                "services": {
                    "description": "must be an array of objectIds from the services collection and is optional",
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                },
                "taxid": {
                    "description": "must be a string and is optional",
                    "type": "string"
                },
                "users": {
                    "description": "must be an array of objectIds from the users collection and is optional",
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                }
            },
            "required": [
                "name",
                "revenue"
            ],
            "title": "Lab"
        },
...
      }
    }
}